ai_agent/requirements.txt
fastapi
uvicorn
pymongo
langchain
langchain-community
langchain-openai
python-dotenv
